renamed everything referring to local scene for the scene debug console; Added app args check for enabling it on prod with the correct parameter
