+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Diagnostics/DiagnosticsContainer.cs
@@ -28,11 +28,11 @@ namespace DCL.Diagnostics
             Sentry?.AddScopeConfigurator(configureScope);
         }
 
-        public static DiagnosticsContainer Create(IReportsHandlingSettings settings, bool enableLocalSceneReporting = false, params (ReportHandler, IReportHandler)[] additionalHandlers)
+        public static DiagnosticsContainer Create(IReportsHandlingSettings settings, bool enableSceneDebugConsole = false, params (ReportHandler, IReportHandler)[] additionalHandlers)
         {
             settings.NotifyErrorDebugLogDisabled();
 
-            int handlersCount = DEFAULT_REPORT_HANDLERS_COUNT + additionalHandlers.Length + (enableLocalSceneReporting ? 1 : 0);
+            int handlersCount = DEFAULT_REPORT_HANDLERS_COUNT + additionalHandlers.Length + (enableSceneDebugConsole ? 1 : 0);
             List<(ReportHandler, IReportHandler)> handlers = new List<(ReportHandler, IReportHandler)>(handlersCount);
             handlers.AddRange(additionalHandlers);
 
@@ -44,7 +44,7 @@ namespace DCL.Diagnostics
             if (settings.IsEnabled(ReportHandler.Sentry))
                 handlers.Add((ReportHandler.Sentry, sentryReportHandler = new SentryReportHandler(settings.GetMatrix(ReportHandler.Sentry), settings.DebounceEnabled)));
 
-            if (enableLocalSceneReporting)
+            if (enableSceneDebugConsole)
                 AddLocalSceneReportingHandler(handlers);
 
             var logger = new ReportHubLogger(handlers);
@@ -55,7 +55,7 @@ namespace DCL.Diagnostics
             Debug.unityLogger.logHandler = logger;
 
             // Enable Hub static accessors
-            ReportHub.Initialize(logger, enableLocalSceneReporting);
+            ReportHub.Initialize(logger, enableSceneDebugConsole);
 
             return new DiagnosticsContainer { ReportHubLogger = logger, defaultLogHandler = defaultLogHandler, Sentry = sentryReportHandler };
         }
@@ -68,7 +68,7 @@ namespace DCL.Diagnostics
             entries.Add(new CategorySeverityMatrix.Entry() { Category = ReportCategory.JAVASCRIPT, Severity = LogType.Exception });
             entries.Add(new CategorySeverityMatrix.Entry() { Category = ReportCategory.JAVASCRIPT, Severity = LogType.Log });
             jsOnlyMatrix.entries = entries;
-            handlers.Add((ReportHandler.DebugLog, new LocalSceneDevelopmentReportHandler(jsOnlyMatrix, false)));
+            handlers.Add((ReportHandler.DebugLog, new SceneDedicatedReportHandler(jsOnlyMatrix, false)));
         }
     }
 }
