+++ Explorer/Assets/Plugins/CommandTerminal/Resources/SceneDebugTerminal.prefab.meta
