+++ Explorer/Assets/Plugins/CommandTerminal/SceneDebugTerminal.asmdef
